Added base logic for piles (tableau, foundation)
Classes 'Tableau' and 'Foundation' extend abstract class CardPile.
Currently implemented logic includes adding cards to a pile (which will
be redundant as we progress - just there now for testing purposes) and
moving card from one pile to another.

The 'moveCard' method is overloaded to handle foundations or tableaus.
To move a card from tableau 1 to foundation 1, you would call the
movecard method from tableau 1:

t1.moveCard(f1);

Tomorrow I will add the logic to handle checking if a card is even able
to be moved to the target pile (ie; it is the opposite colour and one
rank below for tableaus, and is the same suit and one rank above for
foundations.)