Nearly complete card and pile functionality.
Logic is complete to handle cards moving from one pile to another,
with restrictions in place to determine whether target pile can
accept given card. Most logic is handled in CardPile.java, so if a
class looks a little empty, it's because it didn't need many/any
overrides from the superclass.