Updated getImage() method of Card class to return image location
Context: getImage() returns the activeImage String, which is the
location of the image file for a given card. When a card is face down,
it returns "/CardImages/cardBack.svg". If it is face up, it returns the
following:

"/CardImages/" + imageName + ".svg"

So, if the card is the 9 of Hearts, it will return "/CardImages/9H.svg".
The imageName variable is a concatenation of a card's rank (9) and suit
(H).